+++ server/src/domain/system-config/dto/index.ts
@@ -1,5 +1,4 @@
 export * from './system-config-ffmpeg.dto';
-export * from './system-config-library-scan.dto';
 export * from './system-config-library.dto';
 export * from './system-config-oauth.dto';
 export * from './system-config-password-login.dto';
