+++ server/src/domain/system-config/dto/system-config-library.dto.ts
@@ -1,6 +1,42 @@
 import { Type } from 'class-transformer';
-import { IsObject, ValidateNested } from 'class-validator';
-import { SystemConfigLibraryScanDto } from './system-config-library-scan.dto';
+import {
+  IsBoolean,
+  IsNotEmpty,
+  IsObject,
+  IsString,
+  Validate,
+  ValidateIf,
+  ValidateNested,
+  ValidatorConstraint,
+  ValidatorConstraintInterface,
+} from 'class-validator';
+import { CronJob } from 'cron';
+
+const isEnabled = (config: SystemConfigLibraryScanDto) => config.enabled;
+
+@ValidatorConstraint({ name: 'cronValidator' })
+class CronValidator implements ValidatorConstraintInterface {
+  validate(expression: string): boolean {
+    try {
+      new CronJob(expression, () => {});
+    } catch (error) {
+      return false;
+    }
+
+    return true;
+  }
+}
+
+export class SystemConfigLibraryScanDto {
+  @IsBoolean()
+  enabled!: boolean;
+
+  @ValidateIf(isEnabled)
+  @IsNotEmpty()
+  @Validate(CronValidator, { message: 'Invalid cron expression' })
+  @IsString()
+  cronExpression!: string;
+}
 
 export class SystemConfigLibraryDto {
   @Type(() => SystemConfigLibraryScanDto)
