+++ server/src/domain/system-config/system-config.core.ts
@@ -13,6 +13,7 @@ import {
   VideoCodec,
 } from '@app/infra/entities';
 import { BadRequestException, ForbiddenException, Injectable, Logger } from '@nestjs/common';
+import { CronExpression } from '@nestjs/schedule';
 import { plainToInstance } from 'class-transformer';
 import { validate } from 'class-validator';
 import * as _ from 'lodash';
@@ -123,7 +124,7 @@ export const defaults = Object.freeze<SystemConfig>({
   library: {
     scan: {
       enabled: true,
-      cronExpression: '0 0 * * *',
+      cronExpression: CronExpression.EVERY_DAY_AT_MIDNIGHT,
     },
   },
 });
