+++ web/src/routes/admin/system-settings/+page.svelte
@@ -20,8 +20,8 @@
   import Icon from '$lib/components/elements/icon.svelte';
   import type { PageData } from './$types';
   import NewVersionCheckSettings from '$lib/components/admin-page/settings/new-version-check-settings/new-version-check-settings.svelte';
+  import LibrarySettings from '$lib/components/admin-page/settings/library-settings/library-settings.svelte';
   import { mdiAlert, mdiContentCopy, mdiDownload } from '@mdi/js';
-  import LibraryScanSettings from '$lib/components/admin-page/settings/library-scan/library-scan-settings.svelte';
 
   export let data: PageData;
 
@@ -71,7 +71,7 @@
       </SettingAccordion>
 
       <SettingAccordion title="Library" subtitle="Manage library settings">
-        <LibraryScanSettings disabled={$featureFlags.configFile} libraryConfig={configs.library} />
+        <LibrarySettings disabled={$featureFlags.configFile} libraryConfig={configs.library} />
       </SettingAccordion>
 
       <SettingAccordion title="Machine Learning Settings" subtitle="Manage machine learning features and settings">
